moved models to test packages bc is only used there
